+++ test/integration/test_deploy.py
@@ -1,15 +1,17 @@
 import os
+import sys
 import unittest
 
 from . import assertPopen, assertOutput
 from .. import *
 from mike import git_utils, versions
+from mike.commands import AliasType
 
 
 class DeployTestCase(unittest.TestCase):
     def _test_deploy(self, expected_message=None,
                      expected_versions=[versions.VersionInfo('1.0')],
-                     redirect=True, directory='.'):
+                     alias_type=AliasType.symlink, directory='.'):
         rev = git_utils.get_latest_commit('master', short=True)
         message = assertPopen(['git', 'log', '-1', '--pretty=%B']).rstrip()
         if expected_message:
@@ -28,8 +30,10 @@ class DeployTestCase(unittest.TestCase):
                       v_str + '/css/version-select.css',
                       v_str + '/js/version-select.js'}
             for a in v.aliases:
-                files |= {a, a + '/index.html'}
-                if not redirect:
+                files.add(a)
+                if alias_type != AliasType.symlink:
+                    files |= {a + '/index.html'}
+                if alias_type == AliasType.copy:
                     files |= {a + '/css/version-select.css',
                               a + '/js/version-select.js'}
         assertDirectory(directory, files, allow_extra=True)
@@ -59,33 +63,47 @@ class TestDeploy(DeployTestCase):
             versions.VersionInfo('1.0', '1.0.0')
         ])
 
+    @unittest.skipIf(sys.platform == 'win32' and sys.version_info < (3, 8),
+                     'this version of realpath fails to resolve symlinks')
     def test_aliases(self):
         assertPopen(['mike', 'deploy', '1.0', 'latest'])
         check_call_silent(['git', 'checkout', 'gh-pages'])
         self._test_deploy(expected_versions=[
             versions.VersionInfo('1.0', aliases=['latest'])
         ])
+        self.assertTrue(os.path.islink('latest'))
+        self.assertEqual(os.path.normcase(os.path.realpath('latest')),
+                         os.path.normcase(os.path.abspath('1.0')))
+
+    def test_aliases_redirect(self):
+        assertPopen(['mike', 'deploy', '1.0', 'latest',
+                     '--alias-type=redirect'])
+        check_call_silent(['git', 'checkout', 'gh-pages'])
+        self._test_deploy(expected_versions=[
+            versions.VersionInfo('1.0', aliases=['latest'])
+        ], alias_type=AliasType.redirect)
         with open('latest/index.html') as f:
             self.assertRegex(f.read(), match_redir('../1.0/'))
+
+    def test_aliases_custom_redirect(self):
+        assertPopen(['mike', 'deploy', '1.0', 'latest',
+                     '--alias-type=redirect', '-T',
+                     os.path.join(test_data_dir, 'template.html')])
+        check_call_silent(['git', 'checkout', 'gh-pages'])
+        self._test_deploy(expected_versions=[
+            versions.VersionInfo('1.0', aliases=['latest'])
+        ], alias_type=AliasType.redirect)
+        check_call_silent(['git', 'checkout', 'gh-pages'])
+
+        with open('latest/index.html') as f:
+            self.assertEqual(f.read(), 'Redirecting to ../1.0/\n')
 
     def test_aliases_copy(self):
         assertPopen(['mike', 'deploy', '1.0', 'latest', '--alias-type=copy'])
         check_call_silent(['git', 'checkout', 'gh-pages'])
         self._test_deploy(expected_versions=[
             versions.VersionInfo('1.0', aliases=['latest'])
-        ], redirect=False)
-
-    def test_aliases_custom_redirect(self):
-        assertPopen(['mike', 'deploy', '1.0', 'latest', '-T',
-                     os.path.join(test_data_dir, 'template.html')])
-        check_call_silent(['git', 'checkout', 'gh-pages'])
-        self._test_deploy(expected_versions=[
-            versions.VersionInfo('1.0', aliases=['latest'])
-        ])
-        check_call_silent(['git', 'checkout', 'gh-pages'])
-
-        with open('latest/index.html') as f:
-            self.assertEqual(f.read(), 'Redirecting to ../1.0/\n')
+        ], alias_type=AliasType.copy)
 
     def test_update(self):
         assertPopen(['mike', 'deploy', '1.0', 'latest'])
@@ -346,8 +364,9 @@ class TestDeployNoDirectoryUrls(unittest.TestCase):
         check_call_silent(['git', 'add', 'mkdocs.yml', 'docs'])
         check_call_silent(['git', 'commit', '-m', 'initial commit'])
 
-    def test_default(self):
-        assertPopen(['mike', 'deploy', '1.0', 'latest'])
+    def test_alias_redirect(self):
+        assertPopen(['mike', 'deploy', '1.0', 'latest',
+                     '--alias-type=redirect'])
         check_call_silent(['git', 'checkout', 'gh-pages'])
 
         with open('latest/index.html') as f: